master

+++ Entities/Concrete/Car.cs
@@ -10,7 +10,7 @@ namespace Entities.Concrete
         public int ID { get; set; }
         public int BrandID { get; set; }
         public int ModelYear { get; set; }
-        public int DailyPrice { get; set; }
+        public decimal DailyPrice { get; set; }
         public string Description { get; set; }
 
     }
